Feature/kita management

+++ app/Http/Controllers/DomainsController.php
@@ -10,6 +10,7 @@ use App\Http\Requests\Domains\CreateDomainRequest;
 use App\Http\Requests\Domains\ReorderDomainsRequest;
 use App\Http\Requests\Domains\UpdateDomainRequest;
 use App\Models\Domain;
+use App\Models\User;
 use App\Services\Items\DomainItemService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
@@ -44,7 +45,7 @@ class DomainsController extends BaseController
      */
     public function index(Request $request)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $args = $request->only(['page', 'per_page', 'sort', 'order_by', 'search']);
 
@@ -59,12 +60,12 @@ class DomainsController extends BaseController
 
     /**
      * @param Request $request
-     * @param Domain $domain
+     * @param Domain  $domain
      * @return \Inertia\Response
      */
     public function show(Request $request, Domain $domain)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         return Inertia::render('Domains/Partials/ManageDomain', [
             'domain' => $domain->loadMissing(['subdomains' => function ($query) {
@@ -79,10 +80,10 @@ class DomainsController extends BaseController
      */
     public function store(CreateDomainRequest $request)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $attributes = $request->validated();
-        $result = $this->domainItemService->create($attributes);
+        $result     = $this->domainItemService->create($attributes);
 
         return $result
             ? Redirect::back()->withSuccesses(__('crud.domains.create_success'))
@@ -91,15 +92,15 @@ class DomainsController extends BaseController
 
     /**
      * @param UpdateDomainRequest $request
-     * @param Domain $domain
+     * @param Domain              $domain
      * @return \Illuminate\Http\RedirectResponse
      */
     public function update(UpdateDomainRequest $request, Domain $domain)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $attributes = $request->validated();
-        $result = $this->domainItemService->update($domain->id, $attributes);
+        $result     = $this->domainItemService->update($domain->id, $attributes);
 
         return $result
             ? Redirect::back()->withSuccesses(__('crud.domains.update_success'))
@@ -108,12 +109,12 @@ class DomainsController extends BaseController
 
     /**
      * @param Request $request
-     * @param Domain $domain
+     * @param Domain  $domain
      * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Request $request, Domain $domain)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $result = $this->domainItemService->delete($domain->id);
 
@@ -124,12 +125,12 @@ class DomainsController extends BaseController
 
     /**
      * @param Request $request
-     * @param Domain $domain
+     * @param Domain  $domain
      * @return \Illuminate\Http\RedirectResponse
      */
     public function restore(Request $request, Domain $domain)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $result = $this->domainItemService->update($domain->id, [
             'deleted_at' => null,
@@ -146,10 +147,10 @@ class DomainsController extends BaseController
      */
     public function reorder(ReorderDomainsRequest $request)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $attributes = $request->validated();
-        $result = $this->domainItemService->reorder($attributes);
+        $result     = $this->domainItemService->reorder($attributes);
 
         return $result
             ? Redirect::back()->withSuccesses(__('crud.domains.reorder_success'))
